Simple Interest Calculator
Final Project Overview
Now that you have been equipped with the skills to use HTML, CSS, and JavaScript, you will have the opportunity to practice and create a web application.

In this scenario, you are a web developer for an NGO. Everyone Can Get Rich is a non-governmental charity organization. This NGO works to improve the financial literacy of the common people. They have hired a developer to create an online application. This app titled Simple Interest Calculator, will compute the interest one can earn given the amount you wish to save, the number of years you can keep it with the bank and the interest rate the bank would pay you.

The NGO is very particular about the way the app looks including the colors and the alignment of the form elements.

This project will be graded by your peers who are also completing the course during the same session. This project is worth 20 marks of your total grade, and is distributed as follows:

Review Criteria
Basic Functionality (6 marks)

Principal Amount input box validation (3 marks )

How close is the app to the one given in the specification visually? (11 marks)

Overview of Assignment Instructions
Step A: Download the initial project
Download the project in the Theia environment.
Step B: Finish coding the project
Complete the project as per requirements.

Share the relevant screenshots of the app.

Share the html, css and js source files

Submit your work for grading
Upload the relevant screenshots(wherever applicable) of your project when prompted in the peer review section of the final project, you will be graded as below:

The image of your fully designed calculator. (1 mark)
The ‘Principal input box’ is present. (1 mark)
The ‘Interest Slider’ is present. (1 mark)
The ‘dropdown’ box expands to select the number of years. (1 mark)
The ‘compute interest’ button is present. (1 mark)
Check the answer to the values inserted: 1000 as Amount, 10% as Interest rate, and 10 as the number of years, the interest amount should be 1000 after compute interest is clicked. (1 mark)
Check the answer to the values inserted: 4800 as Amount, 15.25% as Interest rate, and 5 as the number of years, the interest amount should be 3660 after compute interest is clicked. (1 mark)
Alert box appears in the screenshot which asks the user to ‘Enter a positive number’ when the value of the principal is entered as zero. (1 mark)
Alert box appears in the screenshot which asks the user to ‘Enter a positive number’ when the value of the principal is entered as a negative number. (1 mark)
If the principal amount field is left blank, and while clicking on “Compute Interest”, there is an alert asking you to enter a positive number. (1 mark)
All the input controls are aligned at the center. (1 mark)
All the colors as specified in the requirements. (1 mark)
The borders of the div that contain the input elements are rounded. (1 mark)
After the Compute Interest is clicked, all the numbers in the output are highlighted in yellow. (1 mark)
Everything should be spelled correctly. (1 mark)
The title “Web App – Simple Interest Calculator” is visible on the browser page/tab. (1 mark)
When the interest rate slider is moved to extreme left, the interest rate should show as 1%. (1 mark)
When the interest rate slider is moved to extreme right, the interest rate should show as 20%. (1 mark)
The slider can be moved to show the value as 4.75%. (1 mark)
At the bottom of the calculator, there is a copyright symbol and the name of the owner. (1 mark)
© IBM Corporation 2022. All rights reserved.